Added Subdirectories for models
Added subdirectories for Models with DTO and Entities for better clarity
