+++ backend/src/main/java/com/fmc/starterApp/controllers/User2Controller.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.fmc.starterApp.models.User2;
+import com.fmc.starterApp.models.entity.User2;
 import com.fmc.starterApp.services.User2Service;
 
 import lombok.AllArgsConstructor;
