--- backend/src/main/java/com/fmc/starterApp/models/PostgresExampleObject.java
@@ -1,23 +1,0 @@
-package com.fmc.starterApp.models;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "example")
-public class PostgresExampleObject {
-
-    @Getter
-    @Setter
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Setter
-    @Getter
-    private String name;
-
-
-}
