+++ backend/src/main/java/com/fmc/starterApp/models/dto/AdminDTO.java
@@ -1,7 +1,8 @@
-package com.fmc.starterApp.models;
+package com.fmc.starterApp.models.dto;
 
 import lombok.Builder;
 import lombok.Data;
+import com.fmc.starterApp.models.entity.AppUser;
 
 import java.util.List;
