+++ backend/src/main/java/com/fmc/starterApp/models/dto/User2DTO.java
@@ -1,9 +1,11 @@
-package com.fmc.starterApp.models;
+package com.fmc.starterApp.models.dto;
 
 import lombok.Builder;
 import lombok.Data;
 
 import java.util.List;
+
+import com.fmc.starterApp.models.entity.User2;
 
 @Builder
 @Data
