+++ backend/src/main/java/com/fmc/starterApp/models/entity/MatrixImage.java
@@ -1,9 +1,19 @@
-package com.fmc.starterApp.models;
-
-import jakarta.persistence.*;
-import lombok.*;
+package com.fmc.starterApp.models.entity;
 
 import java.time.LocalDateTime;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
 @Setter
