+++ backend/src/main/java/com/fmc/starterApp/repositories/CarverMatrixRepository.java
@@ -1,10 +1,11 @@
 package com.fmc.starterApp.repositories;
 
-import com.fmc.starterApp.models.CarverMatrix;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
+import com.fmc.starterApp.models.entity.CarverMatrix;
 
 import java.util.List;
