+++ backend/src/main/java/com/fmc/starterApp/repositories/MatrixImageRepository.java
@@ -1,8 +1,9 @@
 package com.fmc.starterApp.repositories;
 
-import com.fmc.starterApp.models.MatrixImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.fmc.starterApp.models.entity.MatrixImage;
 
 @Repository
 public interface MatrixImageRepository extends JpaRepository<MatrixImage, Long> {
