+++ backend/src/main/java/com/fmc/starterApp/repositories/User2Repository.java
@@ -2,7 +2,7 @@ package com.fmc.starterApp.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.fmc.starterApp.models.User2;
+import com.fmc.starterApp.models.entity.User2;
 
 public interface User2Repository extends JpaRepository<User2, Long> {
 }
