+++ backend/src/main/java/com/fmc/starterApp/repositories/UserLogsRepository.java
@@ -1,7 +1,8 @@
 package com.fmc.starterApp.repositories;
 
-import com.fmc.starterApp.models.UserLogs;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.fmc.starterApp.models.entity.UserLogs;
 
 public interface UserLogsRepository extends JpaRepository<UserLogs, Long> {
