+++ backend/src/main/java/com/fmc/starterApp/repositories/UsersRepository.java
@@ -1,7 +1,8 @@
 package com.fmc.starterApp.repositories;
 
-import com.fmc.starterApp.models.AppUser;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.fmc.starterApp.models.entity.AppUser;
 
 public interface UsersRepository extends JpaRepository<AppUser, Long> {
