+++ backend/src/main/java/com/fmc/starterApp/services/PostGresExampleService.java
@@ -1,6 +1,6 @@
 package com.fmc.starterApp.services;
 
-import com.fmc.starterApp.models.PostgresExampleObject;
+import com.fmc.starterApp.models.entity.PostgresExampleObject;
 import com.fmc.starterApp.repositories.PostgresRepository;
 import org.springframework.transaction.annotation.Transactional;
